Initial Commit.
GeoNames is a big, useful data set found at www.geonames.org.

The file is quite large. Many gigabytes when uncompressed.

It's not an easy-to-use format, and the single giant file is
problematic.

This module is a DFDL schema that parses this data, and that can create
a rational XML representation.